resolved circular dependencies of ongoing games by extracting notifier and provider into ongoing_games_notifier
